requirements.txt
Django==5.0.1
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-environ==0.11.2
PyJWT==2.8.0
psycopg2-binary==2.9.9
Pillow==10.1.0